Move AbstractReservable
